[Pantalla] - Mostrar productos

+++ db/schema.rb
@@ -69,9 +69,9 @@ ActiveRecord::Schema.define(version: 20160320023014) do
     t.string   "description"
     t.float    "price"
     t.string   "image"
-    t.integer  "stock",          default: 0
-    t.integer  "stock_min",      default: 0
-    t.integer  "stock_max",      default: 0
+    t.float    "stock",          default: 0.0
+    t.float    "stock_min",      default: 0.0
+    t.float    "stock_max",      default: 0.0
     t.integer  "measuring_type", default: 0
     t.boolean  "active",         default: false
     t.datetime "created_at"
